Added support for symfony/dot-env while bootstrapping tests
This allows to persist the BBB URL and secret used for integration testing in a .env file and releases the requirement to type an `export` statement every time.

This also eases the usage of the PHPUnit runners provided by IDEs e.g. PhpStorm as it releases the requirement to add special configuration to the PHPUnit runner.